add midi parser into perftests suite
